Add new "init" metapackage
This new package takes over the role of the sysvinit package which was
introduced in [1] and ensures that an init system is installed at all
times.
The complete transition plan is outlined at [2].
For now we keep sysvinit-core as first alternative until we have the
necessary updates for sysvinit in place.

[1] https://packages.qa.debian.org/s/sysvinit/news/20131227T000039Z.html
[2] https://lists.debian.org/debian-devel/2014/07/msg00611.html