Only pod2man init-system-helpers script for now
The new service and *-rc.d have no pod (yet).
Instead their manually written roff in man8/ is shipped.
In the longer run they might want to be rewritten in pod
and then this commit can be reverted.
(Please note though, that invoke-rc.d is a shell script, not perl.)