update-rc.d: Ensure insserv doesn't fail if initscripts is not installed
At some point in the future, it should be possible to install a system
without initscripts package. This means that insserv should ignore any
missing Requires-* package, because that information will be discarded
in a non-initscripts system; all we care about is the enable/disable
link. So pass -f to insserv, so that it still generates the link, even
when some dependencies are missing.

Closes: #806438