invoke-rc.d: Fix check for failing "runlevel" command
The "test" command was just testing the result of sed which never
(realistically) fails.

Note that this changes behaviour in early boot (rcS in SysV, before
sysvinit.target in systemd), as invoke-rc.d will now fail because runlevel
fails with code 1 and says "unknown". This should mostly be considered as a
feature as running things like "smbd reload" from if-up.d/ hooks isn't
safe/working anyway.

This currently breaks invoke-rc.d in late boot under systemd (between
sysvinit.target and default.target) due to LP: #1575572. This will be fixed in
a subsequent commit.

Also replace the sed call with a simple variable substitution.

Spotted as a side issue in <https://launchpad.net/bugs/1575572>.