invoke-rc.d: Fix runlevel check during late boot under systemd
Under systemd, "runlevel" only switches to 3 or 5 when
multi-user.target/graphical.target have been reached, not before. Adjust the
runlevel check accordingly. This is only relevant for the check for
wrong/dangling rcN.d/ symlinks, so just pin it to "5" (the precise value does
not matter much).

Note that this does not affect runlevels 0 and 6, runlevel succeeds there
under systemd. This is only a fallback if runlevel fails (and says "unknown").

Fixing this in systemd requires bigger architectural changes, so use this tiny
(and backportable) bandaid for the time being.

LP: #1575572