invoke-rc.d: Quiesce error message if "runlevel" does not exist
This is now the case in chroots since init stopped being essential.

In this case we fall through to the systemd check, and if that is not running
either just keep $RL as empty. We cannot do the rc?.d symlink checks then, but
in those environments we *really* expect having a policy-rc.d which disables
the starting of services anyway.

Closes: #827376