Keep an internal copy of xsettings-client
xsettings-client code was meant to be a copy-and-paste module, not a
shared library. Grab a copy from
git://anongit.freedesktop.org/xdg/xdg-specs and include it in the
sources.

Remove the build complications: always build with xsettings.

The new files are MIT (old style) licensed: add a COPYING.MIT file.

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>