update-rc.d: preserve S/K links when insserv is not installed
If there are any links, preserve them. In fallback mode we don't care
about the number encoded, but we do care about the S/K state. So instead
of trying to renumber back to 01, preserve the actually existing links.

This also fixes a bug that could cause multiple enable links to be
created for a single service.

Closes: #832945