invoke-rc.d, service: Only ignore systemd unit dependencies before default.target
"systemctl is-system-running" might still be false in case of running
jobs  for device/mount/hotplug/dynamic actions units. But in those cases we
already do want to respect unit dependencies, as the system is booted up
sufficiently to avoid dependency loops. Thus weaken the condition to
"default.target is active".

This does not change the behaviour for single-user: is-system-running has
always been false there, so dependencies continue to be ignored.

Fixes installation of packages like PostgreSQL under cloud-init or when
manually installing packages right after booting.

LP: #1576692