Support running with child namespaces
By default killall should not kill processes in a namespace it is not
part of. If this is allowed, it allows callers to break namespaces they
did not expect to affect, requiring rewrite of all callers to fix.

So by default, we should work in the current namespace. If -N 0 is
specified, then we look at all namespaces, and if any other pid is
specified we look in only that namespace.

Signed-off-by: Debabrata Banerjee <dbanerje@akamai.com>