killall: fix reporting when >32 names specified
Recording which named processes were killed is stored in an unsigned
long bitmask.  Some of the bit-wise math is done with an int.  If
sizeof(long) > sizeof(int), some not-found processes may not be reported
as such, and/or the return code may be wrong.

This is fixed by making sure bit-wise math is being done with a long.