Revert "Drop fallback for old versions of systemd that didn't have is-enabled"
If a package ships both init script and systemd service file, the
systemd unit will not be enabled by the time invoke-rc.d is called
(with current debhelper sequence). This would make systemctl is-enabled
report the wrong status, and then the service would not be started.

Add a comment noting this to avoid removing again in the future.

This reverts commit 6f95680ffc9b1605841eb7d3d8eb92c790e6c73a.

Closes: #906421, #906051