Initial bzip2 test suite.
Contains test files from the commons-compress, dotnetzip, go, lbzip2
and pyflate projects. Each test file has either an associated md5 sum
to check it decompressed correctly. Or it is named bz2.bad to indicate
it cannot be decompressed (or might need --force to decompress).

The run-tests.sh test wrapper runs a bzip2 binary in a couple of
configurations, optionally under valgrind, on all the test files in
the tree.