Remove testfilecopy after second decompression attempt.
We run tests twice, once in default mode and once in small mode.
For the recompression/redecompression attempt we create a copy of
the original input file. Which we remove after the first attempt
in default mode. It should also be removed after the second attempt
in small mode. Otherwise we keep a possibly very large file around.