fuser: Less confused about duplicate dev_id
NFS mounts from the same server have the same device ID. This means
using the -m option a process using one of those mounts will be
"found" in all of the others too.

lsof doesn't have this confusion as it checks the real path against
the mount point and only matches if they start the same.

I think it would be confused with double stacked NFS shares such
as /nfs/SHARE1/blah/SHARE2 with the open file in SHARE2 but
there are limits.

References:
 psmisc/psmisc#10