fuser: Check pathname only on non-block devices
The referenced commit we would check the pathname to
ensure it matched our target. This worked fine for
real files. However for block devices it would fail
because "/dev/sda1" doesn't match "/mnt/myfile".

We only check the pathname if the thing we are matching
against is not a block file.

Thanks to @MarsChan for the report and also the suggested
fix!

References:
 commit 5c979b38253d187a8ecb8e52a0878b8bb668894f
 psmisc/psmisc#31

Signed-off-by: Craig Small <csmall@dropbear.xyz>