invoke-rc.d,service: Stop using ignore-dependencies
Ignoring dependencies during bootup and shutdown can have undesired
side-effects, like changing the ordering of services and ultimately
causing them to fail. So drop this hack, which was added to work around
issues in hook scripts. We only keep a special case for reload requests,
where we schedule those with "no-block" to avoid unnecessary dead-locks.

Closes: #960595