pstree: Don't stop compact with pgids
Showing pgids used to disable compaction. But, despite the vague
NPTL documentation, its not possible to have different PGID in the
same thread group, so lack of compaction is not required.

With setpgid(), the last thread in the process that calls it
wins. There might be a debate about should this be the right way
but that's a kernel thing.

References:
 #34
 Test program https://gitlab.com/-/snippets/2094161
  https://man7.org/linux/man-pages/man7/nptl.7.html