deb-systemd-invoke: add support for starting/stopping units in user instances
Right now if a package adds/removes a user unit, we cannot automatically
start it/stop it/restart it.
systemd v250 adds the ability to talk to user instances of the manager
via the '--machine <ID>@' option. The list of IDs can be queried by
wildcard-querying with 'list-units user@*'.
Use this to add a --user option and let callers start/stop/restart user
units. Print a log message and skip it if systemd < 250.

Mirrors what is done by the RPM scripts:

https://github.com/systemd/systemd/blob/main/src/rpm/systemd-update-helper.in#L53