Do not conflict with obsolete packages in ncurses-term
Some people still use them, and while in general it would be better to
look for supported alternatives, current terminal emulators for the
Linux console are sadly lacking.

It still seems desirable to take over the terminfo files in
ncurses-term, so that users of these terminal emulators have a chance
of finding their $TERM supported on remote systems.

Closes: #994433