killall: use openat and pidfd_send_signal
Using openat and pidfd_send_signal by openig the pid directory
at the start means we cannot have the situation where the PID
has been reused and we filter/kill the wrong process.

In other words, is the open(/proc/1234) the same process as
kill(1234) ? Using pidfd we can be sure.

References:
 psmisc/psmisc#37