fuser: Do not path check sockets
The referenced commit fixed issue #10 where NFS servers have the same ID
and fuser could get confused. That fix was to do a path compare between
the user requested path and the path of the mount.

Block devices were skipped, because /dev/sda2 is not the same as /home
string-wise, even if sda2 was mounted as /home

The new issue is, sockets fail to match when using a mount point. That
is because "socket:[1547]" is not the same as
"/home/user/.gnupg/S.gpg-agent" string-wise, even if named socket 1547
uses that path.

The fix is to skip the path checks when looking at sockets. I'm not sure
if that means we will have confusion with named sockets that have a path
on NFS shares or not but I'm not sure there is a fix for those.

References:
 commit 5c979b38253d187a8ecb8e52a0878b8bb668894f
 psmisc/psmisc#35